Change scaling of query priorities
Rather than treating high as 4x medium and that as 25x low, we now
always use a scaling factor of 1000x between query priorities, and
consider partition transforms to have a priority of zero. In addition,
we now prefer partitions with more related queries over those with fewer
related queries given an otherwise equal priority. This reduces the
conflict potentual between rebuilds, compaction, and queries.

Additionally, this commit adds logs for the number of scheduled
partitions by the index, which helped debug this issue and is generally
worth having.

+++ libvast/test/query_queue.cpp
@@ -80,7 +80,7 @@ uuid make_insert(query_queue& q, system::catalog_lookup_result&& candidates) {
 
 uuid make_insert(query_queue& q, system::catalog_lookup_result&& candidates,
                  uint32_t taste_size,
-                 uint8_t priority = query_context::priority::normal) {
+                 uint64_t priority = query_context::priority::normal) {
   uint32_t cands_size = candidates.size();
   auto query_context = make_random_query_context();
   query_context.priority = priority;
